Move image frame into same module as main window
